+++ allauth/account/tests.py
@@ -4,7 +4,6 @@ import json
 import uuid
 from datetime import timedelta
 
-import django
 from django import forms
 from django.conf import settings
 from django.contrib.auth.models import AbstractUser, AnonymousUser
@@ -12,8 +11,10 @@ from django.contrib.sites.models import Site
 from django.core import mail, validators
 from django.core.exceptions import ValidationError
 from django.db import models
+from django.http import HttpResponseRedirect
 from django.test.client import Client, RequestFactory
 from django.test.utils import override_settings
+from django.urls import reverse
 from django.utils.timezone import now
 
 from allauth.account.forms import BaseSignupForm, SignupForm
@@ -26,10 +27,9 @@ from allauth.tests import Mock, TestCase, patch
 from allauth.utils import get_user_model, get_username_max_length
 
 from . import app_settings
-from ..compat import is_authenticated, reverse
 from .adapter import get_adapter
 from .auth_backends import AuthenticationBackend
-from .signals import user_logged_out
+from .signals import user_logged_in, user_logged_out
 from .utils import (
     filter_users_by_username,
     url_str_to_user_pk,
@@ -135,7 +135,7 @@ class AccountTests(TestCase):
 
     @override_settings(
         ACCOUNT_USERNAME_REQUIRED=True,
-        ACCOUNT_SIGNUP_PASSOWRD_ENTER_TWICE=True)
+        ACCOUNT_SIGNUP_PASSWORD_ENTER_TWICE=True)
     def test_signup_password_twice_form_error(self):
         resp = self.client.post(
             reverse('account_signup'),
@@ -185,7 +185,7 @@ class AccountTests(TestCase):
     def _create_user_and_login(self, usable_password=True):
         password = 'doe' if usable_password else False
         user = self._create_user(password=password)
-        self.client_force_login(user)
+        self.client.force_login(user)
         return user
 
     def test_redirect_when_authenticated(self):
@@ -379,6 +379,26 @@ class AccountTests(TestCase):
         data = json.loads(response.content.decode('utf8'))
         assert 'invalid' in data['form']['errors'][0]
 
+    def test_password_reset_flow_with_email_changed(self):
+        """
+        Test that the password reset token is invalidated if
+        the user email address was changed.
+        """
+        user = self._request_new_password()
+        body = mail.outbox[0].body
+        self.assertGreater(body.find('https://'), 0)
+        EmailAddress.objects.create(
+            user=user,
+            email='other@email.org')
+        # Extract URL for `password_reset_from_key` view
+        url = body[body.find('/password/reset/'):].split()[0]
+        resp = self.client.get(url)
+        self.assertTemplateUsed(
+            resp,
+            'account/password_reset_from_key.%s' %
+            app_settings.TEMPLATE_EXTENSION)
+        self.assertTrue('token_fail' in resp.context_data)
+
     @override_settings(ACCOUNT_LOGIN_ON_PASSWORD_RESET=True)
     def test_password_reset_ACCOUNT_LOGIN_ON_PASSWORD_RESET(self):
         user = self._request_new_password()
@@ -391,7 +411,7 @@ class AccountTests(TestCase):
             resp.url,
             {'password1': 'newpass123',
              'password2': 'newpass123'})
-        self.assertTrue(is_authenticated(user))
+        self.assertTrue(user.is_authenticated)
         # EmailVerificationMethod.MANDATORY sends us to the confirm-email page
         self.assertRedirects(resp, '/confirm-email/')
 
@@ -686,8 +706,6 @@ class AccountTests(TestCase):
         }
     }])
     def test_django_password_validation(self):
-        if django.VERSION < (1, 9, ):
-            return
         resp = self.client.post(
             reverse('account_signup'),
             {'username': 'johndoe',
@@ -804,10 +822,12 @@ class EmailFormTests(TestCase):
             HTTP_X_REQUESTED_WITH='XMLHttpRequest')
         data = json.loads(resp.content.decode('utf8'))
         assert data['data'] == [
-            {'email': 'john1@example.org',
+            {'id': self.email_address.pk,
+             'email': 'john1@example.org',
              'primary': True,
              'verified': True},
-            {'email': 'john2@example.org',
+            {'id': self.email_address2.pk,
+             'email': 'john2@example.org',
              'primary': False,
              'verified': False}]
 
@@ -1016,11 +1036,13 @@ class AuthenticationBackendTests(TestCase):
         backend = AuthenticationBackend()
         self.assertEqual(
             backend.authenticate(
+                request=None,
                 username=user.username,
                 password=user.username).pk,
             user.pk)
         self.assertEqual(
             backend.authenticate(
+                request=None,
                 username=user.email,
                 password=user.username),
             None)
@@ -1032,11 +1054,13 @@ class AuthenticationBackendTests(TestCase):
         backend = AuthenticationBackend()
         self.assertEqual(
             backend.authenticate(
+                request=None,
                 username=user.email,
                 password=user.username).pk,
             user.pk)
         self.assertEqual(
             backend.authenticate(
+                request=None,
                 username=user.username,
                 password=user.username),
             None)
@@ -1048,11 +1072,13 @@ class AuthenticationBackendTests(TestCase):
         backend = AuthenticationBackend()
         self.assertEqual(
             backend.authenticate(
+                request=None,
                 username=user.email,
                 password=user.username).pk,
             user.pk)
         self.assertEqual(
             backend.authenticate(
+                request=None,
                 username=user.username,
                 password=user.username).pk,
             user.pk)
@@ -1075,7 +1101,7 @@ class UtilsTests(TestCase):
         with patch('allauth.account.utils.get_user_model') as mocked_gum:
             mocked_gum.return_value = UUIDUser
             self.assertEqual(url_str_to_user_pk(self.user_id),
-                             self.user_id)
+                             uuid.UUID(self.user_id))
 
     def test_pk_to_url_string_identifies_UUID_as_stringlike(self):
         user = UUIDUser(
@@ -1099,3 +1125,82 @@ class UtilsTests(TestCase):
         self.assertEqual(user_username(user), 'CamelCase')
         # TODO: Actually test something
         filter_users_by_username('camelcase', 'foobar')
+
+
+class ConfirmationViewTests(TestCase):
+    def _create_user(self, username='john', password='doe'):
+        user = get_user_model().objects.create(
+            username=username,
+            is_active=True)
+        if password:
+            user.set_password(password)
+        else:
+            user.set_unusable_password()
+        user.save()
+        return user
+
+    @override_settings(ACCOUNT_EMAIL_CONFIRMATION_HMAC=True,
+                       ACCOUNT_LOGIN_ON_EMAIL_CONFIRMATION=True)
+    def test_login_on_confirm(self):
+        user = self._create_user()
+        email = EmailAddress.objects.create(
+            user=user,
+            email='a@b.com',
+            verified=False,
+            primary=True)
+        key = EmailConfirmationHMAC(email).key
+
+        receiver_mock = Mock()  # we've logged if signal was called
+        user_logged_in.connect(receiver_mock)
+
+        # fake post-signup account_user stash
+        session = self.client.session
+        session['account_user'] = user_pk_to_url_str(user)
+        session.save()
+
+        resp = self.client.post(
+            reverse('account_confirm_email',
+                    args=[key]))
+        email = EmailAddress.objects.get(pk=email.pk)
+        self.assertTrue(email.verified)
+
+        receiver_mock.assert_called_once_with(
+            sender=get_user_model(),
+            request=resp.wsgi_request,
+            response=resp,
+            user=get_user_model().objects.get(username='john'),
+            signal=user_logged_in,
+        )
+
+        user_logged_in.disconnect(receiver_mock)
+
+    @override_settings(ACCOUNT_EMAIL_CONFIRMATION_HMAC=True,
+                       ACCOUNT_LOGIN_ON_EMAIL_CONFIRMATION=True)
+    @patch('allauth.account.views.perform_login')
+    @patch('allauth.account.utils.get_user_model', return_value=UUIDUser)
+    def test_login_on_confirm_uuid_user(self, mocked_gum, mock_perform_login):
+        user = UUIDUser(
+            is_active=True,
+            email='john@example.com',
+            username='john')
+
+        # fake post-signup account_user stash
+        session = self.client.session
+        session['account_user'] = user_pk_to_url_str(user)
+        session.save()
+
+        # fake email and email confirmation to avoid swappable model hell
+        email = Mock(verified=False, user=user)
+        key = 'mockkey'
+        confirmation = Mock(autospec=EmailConfirmationHMAC, key=key)
+        confirmation.email_address = email
+        confirmation.from_key.return_value = confirmation
+        mock_perform_login.return_value = HttpResponseRedirect(redirect_to='/')
+
+        with patch('allauth.account.views.EmailConfirmationHMAC',
+                   confirmation):
+            self.client.post(
+                reverse('account_confirm_email',
+                        args=[key]))
+
+        assert mock_perform_login.called