+++ allauth/account/views.py
@@ -7,13 +7,13 @@ from django.http import (
     HttpResponseRedirect,
 )
 from django.shortcuts import redirect
+from django.urls import reverse, reverse_lazy
 from django.utils.decorators import method_decorator
 from django.views.decorators.debug import sensitive_post_parameters
 from django.views.generic.base import TemplateResponseMixin, TemplateView, View
 from django.views.generic.edit import FormView
 
 from . import app_settings, signals
-from ..compat import is_anonymous, is_authenticated, reverse, reverse_lazy
 from ..exceptions import ImmediateHttpResponse
 from ..utils import get_form_class, get_request_param
 from .adapter import get_adapter
@@ -49,13 +49,14 @@ sensitive_post_parameters_m = method_decorator(
 
 
 def _ajax_response(request, response, form=None, data=None):
-    if request.is_ajax():
+    adapter = get_adapter(request)
+    if adapter.is_ajax(request):
         if (isinstance(response, HttpResponseRedirect) or isinstance(
                 response, HttpResponsePermanentRedirect)):
             redirect_to = response['Location']
         else:
             redirect_to = None
-        response = get_adapter(request).ajax_response(
+        response = adapter.ajax_response(
             request,
             response,
             form=form,
@@ -67,7 +68,7 @@ def _ajax_response(request, response, form=None, data=None):
 class RedirectAuthenticatedUserMixin(object):
 
     def dispatch(self, request, *args, **kwargs):
-        if is_authenticated(request.user) and \
+        if request.user.is_authenticated and \
                 app_settings.AUTHENTICATED_LOGIN_REDIRECTS:
             redirect_to = self.get_authenticated_redirect_url()
             response = HttpResponseRedirect(redirect_to)
@@ -114,7 +115,10 @@ class AjaxCapableProcessFormViewMixin(object):
         return form
 
     def _get_ajax_data_if(self):
-        return self.get_ajax_data() if self.request.is_ajax() else None
+        return (
+            self.get_ajax_data()
+            if get_adapter(self.request).is_ajax(self.request)
+            else None)
 
     def get_ajax_data(self):
         return None
@@ -323,7 +327,7 @@ class ConfirmEmailView(TemplateResponseMixin, View):
         if user_pk_str:
             user_pk = url_str_to_user_pk(user_pk_str)
         user = confirmation.email_address.user
-        if user_pk == user.pk and is_anonymous(self.request.user):
+        if user_pk == user.pk and self.request.user.is_anonymous:
             return perform_login(self.request,
                                  user,
                                  app_settings.EmailVerificationMethod.NONE,
@@ -517,6 +521,7 @@ class EmailView(AjaxCapableProcessFormViewMixin, FormView):
         data = []
         for emailaddress in self.request.user.emailaddress_set.all():
             data.append({
+                'id': emailaddress.pk,
                 'email': emailaddress.email,
                 'verified': emailaddress.verified,
                 'primary': emailaddress.primary,
@@ -759,16 +764,19 @@ class LogoutView(TemplateResponseMixin, View):
     def get(self, *args, **kwargs):
         if app_settings.LOGOUT_ON_GET:
             return self.post(*args, **kwargs)
-        if not is_authenticated(self.request.user):
-            return redirect(self.get_redirect_url())
+        if not self.request.user.is_authenticated:
+            response = redirect(self.get_redirect_url())
+            return _ajax_response(self.request, response)
         ctx = self.get_context_data()
-        return self.render_to_response(ctx)
+        response = self.render_to_response(ctx)
+        return _ajax_response(self.request, response)
 
     def post(self, *args, **kwargs):
         url = self.get_redirect_url()
-        if is_authenticated(self.request.user):
+        if self.request.user.is_authenticated:
             self.logout()
-        return redirect(url)
+        response = redirect(url)
+        return _ajax_response(self.request, response)
 
     def logout(self):
         adapter = get_adapter(self.request)