+++ allauth/socialaccount/providers/draugiem/tests.py
@@ -2,9 +2,9 @@ from hashlib import md5
 
 from django.contrib.auth.models import User
 from django.contrib.sites.models import Site
+from django.urls import reverse
 from django.utils.http import urlencode
 
-from allauth.compat import reverse
 from allauth.socialaccount import providers
 from allauth.socialaccount.models import SocialApp, SocialToken
 from allauth.tests import Mock, TestCase, patch