+++ allauth/socialaccount/providers/facebook/provider.py
@@ -4,12 +4,12 @@ from django.conf import settings
 from django.core.exceptions import ImproperlyConfigured
 from django.middleware.csrf import get_token
 from django.template.loader import render_to_string
+from django.urls import reverse
 from django.utils.crypto import get_random_string
 from django.utils.html import escapejs, mark_safe
 from django.utils.http import urlquote
 
 from allauth.account.models import EmailAddress
-from allauth.compat import reverse
 from allauth.socialaccount.app_settings import QUERY_EMAIL
 from allauth.socialaccount.providers.base import (
     AuthAction,
@@ -23,7 +23,7 @@ from .locale import get_default_locale_callable
 
 
 GRAPH_API_VERSION = getattr(settings, 'SOCIALACCOUNT_PROVIDERS', {}).get(
-    'facebook', {}).get('VERSION', 'v2.5')
+    'facebook', {}).get('VERSION', 'v2.12')
 GRAPH_API_URL = 'https://graph.facebook.com/' + GRAPH_API_VERSION
 
 NONCE_SESSION_KEY = 'allauth_facebook_nonce'