+++ allauth/socialaccount/providers/linkedin_oauth2/provider.py
@@ -1,5 +1,8 @@
 from allauth.socialaccount import app_settings
-from allauth.socialaccount.providers.base import ProviderAccount
+from allauth.socialaccount.providers.base import (
+    ProviderAccount,
+    ProviderException,
+)
 from allauth.socialaccount.providers.oauth2.provider import OAuth2Provider
 
 
@@ -11,7 +14,7 @@ class LinkedInOAuth2Account(ProviderAccount):
         # try to return the higher res picture-urls::(original) first
         try:
             return self.account.extra_data['pictureUrls']['values'][0]
-        except:
+        except Exception:
             # if we can't get higher res for any reason, we'll just return the
             # low res
             pass
@@ -34,6 +37,11 @@ class LinkedInOAuth2Provider(OAuth2Provider):
     account_class = LinkedInOAuth2Account
 
     def extract_uid(self, data):
+        if 'id' not in data:
+            raise ProviderException(
+                'LinkedIn encountered an internal error while logging in. \
+                Please try again.'
+            )
         return str(data['id'])
 
     def get_profile_fields(self):