+++ allauth/socialaccount/providers/naver/provider.py
@@ -1,3 +1,4 @@
+from allauth.account.models import EmailAddress
 from allauth.socialaccount.providers.base import ProviderAccount
 from allauth.socialaccount.providers.oauth2.provider import OAuth2Provider
 
@@ -18,6 +19,17 @@ class NaverProvider(OAuth2Provider):
 
     def extract_uid(self, data):
         return str(data['id'])
+
+    def extract_common_fields(self, data):
+        email = data.get("email")
+        return dict(email=email)
+
+    def extract_email_addresses(self, data):
+        ret = []
+        email = data.get("email")
+        if email:
+            ret.append(EmailAddress(email=email, verified=True, primary=True))
+        return ret
 
 
 provider_classes = [NaverProvider]