+++ allauth/socialaccount/providers/untappd/provider.py
@@ -1,5 +1,6 @@
+from django.urls import reverse
+
 from allauth.account.models import EmailAddress
-from allauth.compat import reverse
 from allauth.socialaccount.providers.base import ProviderAccount
 from allauth.socialaccount.providers.oauth2.provider import OAuth2Provider